feat(calendar): integrate dictionary for multilingual support in calendar components
- Added useDictionary hook to fetch localized strings for calendar features.
- Updated Calendar, UpcomingEvents, and MinifiedUpcomingEvents components to use localized strings for event labels and messages.
- Enhanced EventForm with improved event creation logic and validation.
- Refactored CalendarWeekContainer for better readability and performance.
- Improved UI consistency across timetable components with corrected color handling.
- Added new localization entries for calendar-related strings in both English and Chinese dictionaries.
