+++ src/components/Timetable/TimetableSlotHorizontal.tsx
@@ -74,7 +74,7 @@ const TimetableSlotHorizontal = forwardRef<HTMLDivElement, TimetableSlotProps>(
             <VenueChip
               venue={course.venue}
               color={course.textColor}
-              textColor={course.color}
+              textColor={course.textColor}
             />
           )}
         </div>
